correct the code

+++ 04-DataTypeAndVariables/4-2-01-PracticeIntNum/4-2-01-PracticeIntNum.cs
@@ -13,9 +13,9 @@ namespace _4_2_01_PracticeIntNum
             sbyte num1 = sbyte.Parse(Console.ReadLine());
             byte num2 = byte.Parse(Console.ReadLine());
             short num3 = short.Parse(Console.ReadLine());
-            ushort num4 = ushort.Parse(Console.ReadLine());
-            int num5 = int.Parse(Console.ReadLine());
-            int num6 = int.Parse(Console.ReadLine());
+            int num4 = int.Parse(Console.ReadLine());
+            long num5 = long.Parse(Console.ReadLine());
+            long num6 = long.Parse(Console.ReadLine());
             long num7 = long.Parse(Console.ReadLine());
 
 
@@ -27,11 +27,6 @@ namespace _4_2_01_PracticeIntNum
             Console.WriteLine(num5);
             Console.WriteLine(num6);
             Console.WriteLine(num7);
-
-
-
-
-
         }
     }
 }
